requirements.txt
gradio
requests
Pillow
opencv-python
pandas
openpyxl
pyyaml